fix(ui): repair session URL CI integration
