+++ extensions/clickclack/src/discussions/service.test.ts
@@ -1,9 +1,84 @@
 import { describe, expect, it, vi } from "vitest";
 import { ClickClackHttpError } from "../http-client.js";
+import { controlSessionUrl } from "./control-session-url.js";
 import { fallbackDiscussionLabel } from "./naming.js";
 import { MANAGED_CONTRACT_FIELDS, createHarness, testExternalRef } from "./service-test-support.js";
 
+type SessionUrlContractCase = {
+  sessionKey: string;
+  agentId: string;
+  mainKey: string | undefined;
+  expectedPath: string | null;
+};
+
+// Keep in sync with ui/src/app-navigation.test.ts. Core cannot import plugin internals,
+// and this independently published plugin cannot source-import the workspace contract.
+const SESSION_URL_CONTRACT_CASES = [
+  {
+    sessionKey: "agent:main:main",
+    agentId: "main",
+    mainKey: undefined,
+    expectedPath: "/chat/main",
+  },
+  { sessionKey: "main", agentId: "research", mainKey: undefined, expectedPath: "/chat/research" },
+  {
+    sessionKey: "agent:research:workspace",
+    agentId: "main",
+    mainKey: "workspace",
+    expectedPath: "/chat/research",
+  },
+  {
+    sessionKey: "telegram:12345",
+    agentId: "research",
+    mainKey: undefined,
+    expectedPath: "/chat/research/telegram/12345",
+  },
+  {
+    sessionKey: "agent:main:control-link",
+    agentId: "main",
+    mainKey: undefined,
+    expectedPath: "/chat/main/control-link",
+  },
+  { sessionKey: "agent:main:12345678", agentId: "main", mainKey: undefined, expectedPath: null },
+  {
+    sessionKey: "agent:main:telegram:12345",
+    agentId: "main",
+    mainKey: undefined,
+    expectedPath: "/chat/main/telegram/12345",
+  },
+  {
+    sessionKey: "agent:main:dashboard:12345678-90ab-cdef-1234-567890abcdef",
+    agentId: "main",
+    mainKey: undefined,
+    expectedPath: "/chat/main/12345678",
+  },
+  {
+    sessionKey: "agent:main:dashboard:deadbeef-0aaa-4000-8000-000000000001",
+    agentId: "main",
+    mainKey: "deadbeef",
+    expectedPath: "/chat/main/deadbeef0",
+  },
+  {
+    sessionKey: "agent:main:cron:..:run",
+    agentId: "main",
+    mainKey: undefined,
+    expectedPath: "/chat/main/cron/~dotdot/run",
+  },
+] satisfies readonly SessionUrlContractCase[];
+
 describe("ClickClack discussion service", () => {
+  it("matches the Control UI session URL contract vectors", () => {
+    for (const testCase of SESSION_URL_CONTRACT_CASES) {
+      const url = controlSessionUrl(
+        "https://control.example",
+        testCase.sessionKey,
+        testCase.agentId,
+        testCase.mainKey,
+      );
+      expect(url ? new URL(url).pathname : null).toBe(testCase.expectedPath);
+    }
+  });
+
   it("opens a managed channel once and returns stable info URLs", async () => {
     const harness = createHarness({ label: "Release Planning", category: "Projects" });
     harness.config.channels!.clickclack!.apiBaseUrl = "http://127.0.0.1:8484";
