+++ extensions/tsconfig.package-boundary.paths.json
@@ -743,9 +743,6 @@
       "@openclaw/retry": [
         "../dist/plugin-sdk/packages/retry/src/index.d.ts"
       ],
-      "@openclaw/session-url-contract": [
-        "../packages/session-url-contract/src/index.ts"
-      ],
       "@openclaw/workboard-contract": [
         "../packages/workboard-contract/src/index.ts"
       ],
