+++ ui/src/e2e/session-transcript-search.e2e.test.ts
@@ -5,6 +5,7 @@ import { chromium, type Browser, type BrowserContext, type Page } from "playwrig
 import { afterAll, afterEach, beforeAll, describe, expect, it } from "vitest";
 import {
   canRunPlaywrightChromium,
+  controlUiSessionPath,
   installMockGateway,
   resolvePlaywrightChromiumExecutablePath,
   startControlUiE2eServer,
@@ -178,7 +179,9 @@ describeControlUiE2e("Control UI session transcript search", () => {
     await result.waitFor({ state: "visible", timeout: 10_000 });
     await expect.poll(async () => gateway.getRequests("sessions.search")).toHaveLength(2);
     await result.click();
-    await expect.poll(() => page?.url()).toContain("session=agent%3Amain%3Alaunch");
+    await expect
+      .poll(() => (page ? new URL(page.url()).pathname : ""))
+      .toBe(controlUiSessionPath("agent:main:launch"));
     await page
       .getByText("The nebula launch checklist is ready.", { exact: true })
       .waitFor({ state: "visible", timeout: 10_000 });
